# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'vipFarmaBackEnd'","description":"'Proyecto de Farmacia VIP Back End en SpringBoot'","contact":{"name":"Hernán Jesús Morais","email":"104473@tecnicatura.frc.utn.edu.ar"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'Proyecto de Farmacia VIP Back End en SpringBoot'"}],"paths":{"/stock/orders/{id}":{"get":{"tags":["stock-order-controller"],"operationId":"getOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StockOrder"}}}}}},"put":{"tags":["stock-order-controller"],"operationId":"changeOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockOrder"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StockOrder"}}}}}}},"/stock/orders/{id}/status/{status}":{"put":{"tags":["stock-order-controller"],"operationId":"changeStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"status","in":"path","required":true,"schema":{"type":"string","enum":["CREATED","PENDING_DELIVERY","STORED","RECEIVED","REJECTED","CANCELED"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockOrder"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StockOrder"}}}}}}},"/sale/orders/{id}":{"get":{"tags":["sale-order-controller"],"operationId":"getOrder_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SaleOrder"}}}}}},"put":{"tags":["sale-order-controller"],"operationId":"changeOrder_1","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaleOrder"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SaleOrder"}}}}}}},"/sale/orders/{id}/status/{status}":{"put":{"tags":["sale-order-controller"],"operationId":"changeStatus_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"status","in":"path","required":true,"schema":{"type":"string","enum":["CREATED","READY_TO_BILL","BILLED","DELIVERED","CLOSED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SaleOrder"}}}}}}},"/products/{id}":{"get":{"tags":["product-controller"],"operationId":"getProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}},"put":{"tags":["product-controller"],"operationId":"updateProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}},"delete":{"tags":["product-controller"],"operationId":"deleteById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/warehouses":{"get":{"tags":["warehouse-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Warehouse"}}}}}}},"post":{"tags":["warehouse-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Warehouse"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Warehouse"}}}}}}},"/user":{"post":{"tags":["user-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/user/login":{"post":{"tags":["user-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/stock/orders":{"get":{"tags":["stock-order-controller"],"operationId":"getOrders","parameters":[{"name":"stockOrderStatus","in":"query","required":false,"schema":{"type":"string","enum":["CREATED","PENDING_DELIVERY","STORED","RECEIVED","REJECTED","CANCELED"]}},{"name":"warehouseId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"branchOfficeId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StockOrder"}}}}}}},"post":{"tags":["stock-order-controller"],"operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockOrder"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StockOrder"}}}}}}},"/sale/orders":{"get":{"tags":["sale-order-controller"],"operationId":"getOrders_1","parameters":[{"name":"saleOrderStatus","in":"query","required":false,"schema":{"type":"string","enum":["CREATED","READY_TO_BILL","BILLED","DELIVERED","CLOSED"]}},{"name":"branchOfficeId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SaleOrder"}}}}}}},"post":{"tags":["sale-order-controller"],"operationId":"create_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaleOrder"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SaleOrder"}}}}}}},"/racks":{"get":{"tags":["rack-controller"],"operationId":"getAll_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Rack"}}}}}}},"post":{"tags":["rack-controller"],"operationId":"create_4","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Rack"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Rack"}}}}}}},"/products":{"get":{"tags":["product-controller"],"operationId":"getProducts","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"barcode","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}},"post":{"tags":["product-controller"],"operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}}},"/offers":{"get":{"tags":["offer-controller"],"operationId":"getAll_3","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Offer"}}}}}}},"post":{"tags":["offer-controller"],"operationId":"create_5","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Offer"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Offer"}}}}}}},"/lockers":{"get":{"tags":["locker-controller"],"operationId":"getAll_4","parameters":[{"name":"productId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"availability","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Locker"}}}}}}},"post":{"tags":["locker-controller"],"operationId":"create_6","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Locker"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Locker"}}}}}}},"/healthInsurance":{"get":{"tags":["health-insurance-controller"],"operationId":"getAll_5","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HealthInsurance"}}}}}}},"post":{"tags":["health-insurance-controller"],"operationId":"create_7","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthInsurance"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HealthInsurance"}}}}}}},"/healthInsurancePlan":{"get":{"tags":["health-insurance-plan-controller"],"operationId":"getAll_6","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HealthInsurancePlan"}}}}}}},"post":{"tags":["health-insurance-plan-controller"],"operationId":"create_8","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthInsurancePlan"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HealthInsurancePlan"}}}}}}},"/customers":{"get":{"tags":["customer-controller"],"operationId":"getAll_7","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Customer"}}}}}}},"post":{"tags":["customer-controller"],"operationId":"login_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Customer"}}}}}}},"/branchOffice":{"get":{"tags":["branch-office-controller"],"operationId":"getAll_8","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BranchOffice"}}}}}}},"post":{"tags":["branch-office-controller"],"operationId":"create_9","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BranchOffice"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BranchOffice"}}}}}}},"/bills":{"get":{"tags":["bill-controller"],"operationId":"getAll_9","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Bill"}}}}}}},"post":{"tags":["bill-controller"],"operationId":"create_10","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Bill"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Bill"}}}}}}},"/billing/order/{id}":{"post":{"tags":["bill-controller"],"operationId":"billOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"preview","in":"query","required":true,"schema":{"type":"boolean"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogged"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Bill"}}}}}}},"/warehouses/{id}":{"get":{"tags":["warehouse-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Warehouse"}}}}}}},"/users":{"get":{"tags":["user-controller"],"operationId":"getAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}}},"/user/{userId}":{"get":{"tags":["user-controller"],"operationId":"getUserById","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/stocks/offers":{"get":{"tags":["offer-controller"],"operationId":"getAllWithStock","parameters":[{"name":"branchOfficeId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"productId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OfferStock"}}}}}}}},"/stock/summary":{"get":{"tags":["stock-controller"],"operationId":"getStocks","parameters":[{"name":"productId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"stockStatus","in":"query","required":false,"schema":{"type":"string","enum":["ACTIVE","INACTIVE","EXPIRED","RESERVED"]}},{"name":"lockerId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StockSummary"}}}}}}}},"/stock/offer":{"get":{"tags":["offer-controller"],"operationId":"getOfferByParameters","parameters":[{"name":"branchOfficeId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"productId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"planId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OfferStock"}}}}}}},"/racks/{id}":{"get":{"tags":["rack-controller"],"operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Rack"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/offers/{id}":{"get":{"tags":["offer-controller"],"operationId":"getById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Offer"}}}}}}},"/offers/{id}/stock":{"get":{"tags":["offer-controller"],"operationId":"getOfferStock","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Offer"}}}}}}},"/lockers/{id}":{"get":{"tags":["locker-controller"],"operationId":"getById_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Locker"}}}}}}},"/healthInsurancePlan/{id}":{"get":{"tags":["health-insurance-plan-controller"],"operationId":"getById_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HealthInsurancePlan"}}}}}}},"/healthInsurancePlan/insurance/{id}":{"get":{"tags":["health-insurance-plan-controller"],"operationId":"getByInsuranceId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HealthInsurancePlan"}}}}}}}},"/healthInsurance/{id}":{"get":{"tags":["health-insurance-controller"],"operationId":"getById_5","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HealthInsurance"}}}}}}},"/customers/{id}":{"get":{"tags":["customer-controller"],"operationId":"getById_6","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Customer"}}}}}},"delete":{"tags":["customer-controller"],"operationId":"deleteById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/branchOffice/{id}":{"get":{"tags":["branch-office-controller"],"operationId":"getById_7","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BranchOffice"}}}}}}},"/bills/{id}":{"get":{"tags":["bill-controller"],"operationId":"getById_8","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Bill"}}}}}}}},"components":{"schemas":{"BranchOffice":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"location":{"type":"string"}}},"Locker":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"rack":{"$ref":"#/components/schemas/Rack"},"product":{"$ref":"#/components/schemas/Product"},"stockCapacity":{"type":"integer","format":"int32"},"occupiedCapacity":{"type":"integer","format":"int32"},"currentStock":{"type":"integer","format":"int32"},"branchOfficeId":{"type":"integer","format":"int64"}}},"Product":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"barcode":{"type":"integer","format":"int64"},"laboratory":{"type":"string"},"prescriptionRequired":{"type":"boolean"},"price":{"type":"number"}}},"Rack":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"warehouse":{"$ref":"#/components/schemas/Warehouse"},"lockers":{"type":"array","items":{"$ref":"#/components/schemas/Locker"}}}},"StockOrder":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"purchaseOrderId":{"type":"integer","format":"int64"},"warehouse":{"$ref":"#/components/schemas/Warehouse"},"createdDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"stockOrderStatus":{"type":"string","enum":["CREATED","PENDING_DELIVERY","STORED","RECEIVED","REJECTED","CANCELED"]},"statusReason":{"type":"string"},"statusComment":{"type":"string"},"stockOrderItems":{"type":"array","items":{"$ref":"#/components/schemas/StockOrderItem"}}}},"StockOrderItem":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"stockOrderId":{"type":"integer","format":"int64"},"product":{"$ref":"#/components/schemas/Product"},"requiredQuantity":{"type":"integer","format":"int32"},"receivedQuantity":{"type":"integer","format":"int32"},"rejectedQuantity":{"type":"integer","format":"int32"},"actualQuantity":{"type":"integer","format":"int32"},"stockOrderItemStatus":{"type":"string","enum":["CREATED","PENDING","RETURNED","NO_RECEIVED","COMPLETELY_RECEIVED","PARTIALLY_RECEIVED","STORED","CANCELED"]},"lockersToSave":{"type":"array","items":{"$ref":"#/components/schemas/Locker"}}}},"Warehouse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"location":{"type":"string"},"branchOffice":{"$ref":"#/components/schemas/BranchOffice"},"racks":{"type":"array","items":{"$ref":"#/components/schemas/Rack"}}}},"Customer":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"lastName":{"type":"string"},"identificationType":{"type":"string","enum":["DNI","CUIT","CUIL"]},"identification":{"type":"string"},"address":{"type":"string"},"healthInsurance":{"$ref":"#/components/schemas/HealthInsurance"},"healthInsurancePlan":{"$ref":"#/components/schemas/HealthInsurancePlan"}}},"HealthInsurance":{"required":["name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"number":{"type":"string"},"availablePlans":{"type":"array","items":{"$ref":"#/components/schemas/HealthInsurancePlan"}}}},"HealthInsurancePlan":{"required":["healthInsuranceId","name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"healthInsuranceId":{"type":"integer","format":"int64"}}},"Offer":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"product":{"$ref":"#/components/schemas/Product"},"healthInsurance":{"$ref":"#/components/schemas/HealthInsurance"},"healthInsurancePlan":{"$ref":"#/components/schemas/HealthInsurancePlan"},"discountType":{"type":"string","enum":["OVERRIDE","PERCENTAGE","FIXED"]},"discountValue":{"type":"number"},"finalPrice":{"type":"number"}}},"SaleOrder":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/User"},"customer":{"$ref":"#/components/schemas/Customer"},"branchOfficeId":{"$ref":"#/components/schemas/BranchOffice"},"createdDate":{"type":"string","format":"date"},"saleOrderStatus":{"type":"string","enum":["CREATED","READY_TO_BILL","BILLED","DELIVERED","CLOSED"]},"statusDetail":{"type":"string"},"saleOrderItems":{"type":"array","items":{"$ref":"#/components/schemas/SaleOrderItem"}},"totalAmount":{"type":"number"}}},"SaleOrderItem":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"orderId":{"type":"integer","format":"int64"},"offer":{"$ref":"#/components/schemas/Offer"},"quantity":{"type":"integer","format":"int32"},"unitaryPrice":{"type":"number"},"discountAmount":{"type":"number"},"totalPrice":{"type":"number"},"stocks":{"type":"array","items":{"$ref":"#/components/schemas/Stock"}}}},"Stock":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"product":{"$ref":"#/components/schemas/Product"},"createdDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"dueDate":{"type":"string","format":"date"},"initialStock":{"type":"integer","format":"int32"},"availableStock":{"type":"integer","format":"int32"},"locker":{"$ref":"#/components/schemas/Locker"},"stockStatus":{"type":"string","enum":["ACTIVE","INACTIVE","EXPIRED","RESERVED"]}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"userRole":{"type":"string","enum":["ADMIN","CASHIER","SELLER","WAREHOUSE_MANAGER"]},"branchOffice":{"$ref":"#/components/schemas/BranchOffice"}}},"Bill":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/User"},"customer":{"$ref":"#/components/schemas/Customer"},"orderId":{"type":"integer","format":"int64"},"createdDate":{"type":"string","format":"date"},"dueDate":{"type":"string","format":"date"},"items":{"type":"array","items":{"$ref":"#/components/schemas/BillItem"}},"cae":{"type":"string"},"totalAmount":{"type":"number"}}},"BillItem":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"billId":{"type":"integer","format":"int64"},"offer":{"$ref":"#/components/schemas/Offer"},"quantity":{"type":"integer","format":"int32"},"labelInvoice":{"type":"string"},"unitaryPrice":{"type":"number"},"discountAmount":{"type":"number"},"totalPrice":{"type":"number"}}},"UserLogged":{"required":["id"],"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"OfferStock":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"product":{"$ref":"#/components/schemas/Product"},"healthInsurance":{"$ref":"#/components/schemas/HealthInsurance"},"healthInsurancePlan":{"$ref":"#/components/schemas/HealthInsurancePlan"},"discountType":{"type":"string","enum":["OVERRIDE","PERCENTAGE","FIXED"]},"discountValue":{"type":"number"},"finalPrice":{"type":"number"},"stock":{"type":"integer","format":"int64"}}},"StockSummary":{"type":"object","properties":{"stock":{"$ref":"#/components/schemas/Stock"},"value":{"type":"integer","format":"int64"}}}}}}